+++ NEWS.md
@@ -1,3 +1,5 @@
+# mlr3learners (development version)
+
 # mlr3learners 0.15.1
 
 * compatibility: mlr3 1.7.2